Cast toMicroSeconds() to double in cpp98 throughput demo
repetitions * seqSize * wireSize * 8.0 / tm.toMicroSeconds() implicitly
converted the Int64 result of toMicroSeconds() to double; the macos-26
clang flags this as -Wimplicit-int-float-conversion, which -Werror turns
into a hard error. Cast it explicitly.

+++ cpp98/Ice/throughput/Client.cpp
@@ -425,7 +425,7 @@ run(const Ice::CommunicatorPtr& communicator)
                         break;
                     }
                 }
-                double mbit = repetitions * seqSize * wireSize * 8.0 / tm.toMicroSeconds();
+                double mbit = repetitions * seqSize * wireSize * 8.0 / static_cast<double>(tm.toMicroSeconds());
                 if(c == 'e')
                 {
                     mbit *= 2;
